Added portfolio html and css file to github!
Used "TextEdit" on my Mac, chose the "plain text" option, saved the html file as an html file after hitting save for the first time, saved the css file as a "UT-8" file and put ".css" at the end of the name of the document so it read "styles.css" and saved it. Put them in a folder on my desktop called "portfolio" and dragged and dropped here! And really hope this was the correct method.